# openapi.json
{"openapi":"3.0.2","info":{"title":"Aspect4 Order API","description":"API to fetch customer orders from Aspect4","version":"1.0.0"},"paths":{"/orders/{customer_id}":{"get":{"summary":"Get Customer Orders","description":"Fetch orders for a specific customer.","operationId":"GetCustomerOrders","parameters":[{"name":"customer_id","in":"path","required":true,"schema":{"type":"string","description":"The ID of the customer to fetch orders for","title":"Customer Id"},"description":"The ID of the customer to fetch orders for"},{"name":"order_number","in":"query","required":false,"schema":{"description":"Specific order number to fetch","title":"Order Number","type":"string","nullable":true},"description":"Specific order number to fetch"},{"name":"days","in":"query","required":false,"schema":{"type":"integer","description":"Number of days to look back","default":30,"title":"Days"},"description":"Number of days to look back"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Order"},"title":"Response Getcustomerorders"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Order":{"properties":{"order_number":{"type":"string","title":"Order Number"},"order_date":{"type":"string","title":"Order Date"},"order_lines":{"items":{"$ref":"#/components/schemas/OrderLine"},"type":"array","title":"Order Lines"},"status_lines":{"items":{"$ref":"#/components/schemas/OrderLine"},"type":"array","title":"Status Lines"}},"type":"object","required":["order_number","order_date","order_lines","status_lines"],"title":"Order"},"OrderLine":{"properties":{"varenr":{"type":"string","title":"Varenr"},"raw_data":{"type":"string","title":"Raw Data"}},"type":"object","required":["varenr","raw_data"],"title":"OrderLine"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}